flashy bundle + roles to the methods comment

+++ src/Controller/BlogController.php
@@ -8,9 +8,11 @@ use App\Form\CommentType;
 use Cocur\Slugify\Slugify;
 use App\Repository\ArticleRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
@@ -57,12 +59,18 @@ class BlogController extends AbstractController
     }
 
     // ---------------------------------Ajout/Edition d'un commentaire article--------------------------------- //
+    #[IsGranted('ROLE_ADMIN')] 
+    #[IsGranted('ROLE_USER')]
     #[Route('blog/{slug}/comment', name: 'app_article_addComment', methods: ['POST'], requirements: ['slug' => '[a-z0-9\-]*'])]
     #[Route('blog/{slug}/comment/{id}/edit', name: 'app_article_editComment', requirements: ['id' => '\d+'])]
-    public function add_editComment(string $slug, Request $request, Comment $comment = null, ArticleRepository $articleRepository, EntityManagerInterface $entityManager): Response
+    public function add_editComment(string $slug, Request $request, Comment $comment = null, ArticleRepository $articleRepository, EntityManagerInterface $entityManager, Security $security): Response
     {
-        $article = $articleRepository->findOneBy(['slug' => $slug]);
+        $user = $security->getUser();
 
+        if ((!$comment->getUser() === $user) || !$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Veuillez vous connecter ou vous assurer d\'avoir les droits !');
+        }
+        $article = $articleRepository->findOneBy(['slug' => $slug]);
         if (!$article) {
             throw new NotFoundHttpException('Article not found');
         }
@@ -70,12 +78,12 @@ class BlogController extends AbstractController
         if(!$comment){
             $comment = new Comment();
         }
-        
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-
+            // Associer l'utilisateur actuel au commentaire
+            $comment->setUser($user);
             $comment->setArticle($article);
             $entityManager->persist($comment);
             $entityManager->flush();
@@ -90,25 +98,30 @@ class BlogController extends AbstractController
     }
     
     // ---------------------------------Suppression d'un commentaire article--------------------------------- //
+    #[IsGranted('ROLE_ADMIN')] 
+    #[IsGranted('ROLE_USER')]
     #[Route('blog/{slug}/comment/{id}/delete', name: 'app_article_deleteComment', requirements: ['id' => '\d+'])]
-    public function deleteComment(string $slug, int $id, ArticleRepository $articleRepository, EntityManagerInterface $entityManager): Response
+    public function deleteComment(string $slug, int $id, ArticleRepository $articleRepository, EntityManagerInterface $entityManager, Security $security): Response
     {
         $article = $articleRepository->findOneBy(['slug' => $slug]);
-
         if (!$article) {
             throw new NotFoundHttpException('Article not found');
         }
 
+        $user = $security->getUser();
+
         // Recherche le commentaire à supprimer
         $comment = $entityManager->getRepository(Comment::class)->find($id);
-
         if (!$comment) {
             throw new NotFoundHttpException('Comment not found');
         }
-
-        // Supprimez le commentaire
-        $entityManager->remove($comment);
-        $entityManager->flush();
+        if (($comment->getUser() === $user) || $this->isGranted('ROLE_ADMIN')) {
+            // Supprimez le commentaire
+            $entityManager->remove($comment);
+            $entityManager->flush();
+        }else{
+            $this->addFlash('error', 'Veuillez vous connecter ou vous assurer d\'avoir les droits !');
+        }
 
         return $this->redirectToRoute('app_article', ['slug' => $slug]);
     }
